Blocks: In scan_delimiters extract static method (#43984)
Ports 185482-ghe-Automattic/wpcom

When `scan_delimiters()` is attempting to find the end of an HTML
comment, it’s calling a convenience method which is also a closure
in the code (or an anonymous function). This carries a marginal
performance overhead and limits reuse.

In this change that method is extracted onto the class as its own static
method; this should simplify reasoning about the method itself and
provide a perofmrnace lift to the calling code.

Co-authored-by: Dennis Snell <dennis.snell@automattic.com>
Co-authored-by: Brandon Kraft <public@brandonkraft.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/15779890774

Upstream-Ref: Automattic/jetpack@39c83c434badd9c9ade7754ed1062a886dfef4f3